requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
apprise==1.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-lifespan==2.1.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.4
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
coolname==2.2.0
croniter==5.0.1
cryptography==44.0.0
dateparser==1.2.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.15
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
fqdn==1.5.1
fsspec==2024.10.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.0
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.4
jinja2-humanize-extension==0.4.0
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
mako==1.3.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mlb-statsapi==1.8.1
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook-shim==0.2.4
numpy==2.2.0
oauthlib==3.2.2
opentelemetry-api==1.29.0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pip==24.3.1
platformdirs==4.3.6
prefect==3.1.6
prometheus-client==0.21.1
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pygments==2.18.0
pyjwt==2.10.1
pyopenssl==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.12
send2trash==1.8.3
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
soupsieve==2.6
sqlalchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
terminado==0.18.1
text-unidecode==1.3
time-machine==2.16.0
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typer==0.13.1
types-python-dateutil==2.9.0.20241003
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.17.0
zipp==3.21.0
